feat: create generic select component and use it for preferences UI
