+++ apps/client/libs/ui/icons/index.tsx
@@ -1,0 +1,4 @@
+export { FiSun as SunIcon } from "react-icons/fi";
+export { FiMoon as MoonIcon } from "react-icons/fi";
+export { FaAngleDown as ArrowDownIcon } from "react-icons/fa6";
+export { CiGlobe as GlobeIcon } from "react-icons/ci";
