+++ apps/client/package.json
@@ -15,6 +15,7 @@
     "install": "^0.13.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
     "react-use": "^17.6.0",
     "zustand": "^5.0.8"
   },
