+++ apps/client/src/components/Header/components/Navigation/components/Preferences/Preferences.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Flex } from "antd";
+import SelectTheme from "./components/SelectTheme";
+import SelectLanguage from "./components/SelectLanguage";
+
+const Preferences = () => {
+  return (
+    <Flex className="preferences" gap={6}>
+      <SelectTheme />
+      <SelectLanguage />
+    </Flex>
+  );
+};
+
+export default Preferences;
