+++ apps/client/src/components/Header/components/Navigation/components/Preferences/components/SelectTheme.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import useStore from "@/store/store";
+import { MoonIcon, SunIcon } from "@/libs/ui/icons";
+import GenSelect from "@/libs/ui/components/GenSelect";
+
+const SelectTheme = () => {
+  const { theme, setTheme } = useStore();
+  const themeOptions = [
+    {
+      label: <SunIcon />,
+      value: "light",
+    },
+    {
+      label: <MoonIcon />,
+      value: "dark",
+    },
+  ];
+
+  return (
+    <GenSelect
+      options={themeOptions}
+      value={theme}
+      onChange={(value) => setTheme(value)}
+    />
+  );
+};
+
+export default SelectTheme;
